Replace np.float with np.float64 to fix numpy 1.20 deprecation

+++ Corrfunc/utils.py
@@ -138,8 +138,8 @@ def convert_3d_counts_to_cf(ND1, ND2, NR1, NR2,
 
     nonzero = pair_counts['R1R2'] > 0
     if 'LS' in estimator or 'Landy' in estimator:
-        fN1 = np.float(NR1) / np.float(ND1)
-        fN2 = np.float(NR2) / np.float(ND2)
+        fN1 = np.float64(NR1) / np.float64(ND1)
+        fN2 = np.float64(NR2) / np.float64(ND2)
         cf = np.zeros(nbins)
         cf[:] = np.nan
         cf[nonzero] = (fN1 * fN2 * pair_counts['D1D2'][nonzero] -
@@ -793,7 +793,7 @@ def gridlink_sphere(thetamax,
 
     # Upper and lower limits of the declination bands
     grid_dtype= np.dtype({'names':['dec_limit','ra_limit'],
-                          'formats':[(np.float, (2, )), (np.float, (2, ))]
+                          'formats':[(np.float64, (2, )), (np.float64, (2, ))]
     })
     if not link_in_ra:
         sphere_grid = np.zeros(ngrid_dec, dtype=grid_dtype)
